refactor server.js
