+++ client.config.js
@@ -1,0 +1,8 @@
+const { Client, GatewayIntentBits} = require('discord.js');
+
+const client = new Client({
+	intents:
+		[GatewayIntentBits.Guilds]
+});
+
+module.exports = client;
