+++ credentials.json
@@ -1,0 +1,13 @@
+{
+    "type": "service_account",
+    "project_id": "discord-bot-412100",
+    "private_key_id": "fb7108435a3a550ef3a643f6a8824134df5bcedc",
+    "private_key": "-----BEGIN PRIVATE KEY-----\nMIIEvAIBADANBgkqhkiG9w0BAQEFAASCBKYwggSiAgEAAoIBAQDbiitPiVSYVETK\ng0BnXrtTcAv6EzCVDzn8oArGYlaZVdNa1VXzORtbkGpmfEk6p2NDOaTpaiCsnyzy\n2Cv3BX3zy3O4QJVdHhyOXw0pWxnZ3iHBg+N6UvVbtJxRBp3MHaDTEnlmMl+x3+qc\neTM14ioRI2cZe3j/vFWiP6qR0ugDv0V/g7TXwLktJSotlVP+XVxZYr2E9hC1rYOO\nLj3p9eJsRfUxhockWT7RqpLi4YzzunkJsM6Nf7aVojk8aXQ7En/6nmdqcaGztYb3\n6A8qvdeOpTZLo07xosqx/tH/bhlQtZDVnyKuwb5YdJCmVpkVuUik3cpSoTgkO46K\njMt93adFAgMBAAECgf8ZjV7u57L1qRIfwl/wxIxn4p30BtX6z0riRxO3mxlTTbg0\nKuLRXhtRDPSEycMRhx4nuaWP0S3y6Ypy8k3f1gFvWgEiwX4JSk1BdEn8LDFDHCqP\nQ8XJL+gl+HwAR5oynja2NIf2MnT6YDaws8lswzawcqKSFxQMdwnfWImseNGIW4hb\nf2UMfNVEc39Gzni3p3+NtXVQg2ZKpB5cLeaVS+VPsOYHILa7FMaqmbZRrXevgbWZ\nb+qz6gsa/orlLk7rYFjdpSpsThSvnpBilwpqZ6JNHmXQYGDcNbkKkZ2ugrxuy1ma\nATrG/uq6GdwCuGdxQrYXLze0+QL5moH35mKK6mECgYEA88M+jvs858+/xhLgA73o\n6p9YdwKdIOYHBwPPb3usL6HHjeHRIkPV1jAnX9f6SwwHAnicS43SyLTHZFxAWi8c\nRNxpIRUqAkT9/31ZbXay/wveJWnOIypPlyF48yAMO5LIgS5CCAxl/MYjvvRR/P9G\ngoYSqgOm2qPSfITIpvpKedECgYEA5o+eoEqr+QS4FIEbtqs/M1JvC+rORGGtdqTE\nntSmAE3OOLj4DScCKCJ+g8Od1Z3HnhF93AI9qwtUjFFuypY6lVGBiJImfuSk1Y+C\nOPaywJ/zLcft7hSwCr6LXHtUuxC3t8R0WhRo4lfFLadPKtaqGDl2243HzSpMvTz5\nRXb4PzUCgYB4DKv+ltUEBViZICqbLShMhHV5BxyDW0YSulmquemu+6nfwlrQ2T+y\n4LxDg4aM9ztsiCHRQ+OPXRjl4FkBETqG31YXKOkrfPbqoL0uH8s/uISLQHR49RmF\ngMKkWiAsJroAhNuYQyLnPl3UaVcHlmEPXyBKAspl21CKaoq3UYTkQQKBgQCmLI9U\n0lz+ntMaHC/gSa7Ei75zpWnDSW9fmsQIAhYfE/T1VwOZXysjyDDNQTlKFG0iVVMU\nNh1CYWkvtSWYKAzTQCqPR18tiBfxwHxIGC6fvOjlbzripNB4/fqtzko0hYkUjyiZ\nICq0MDRygisCIjP4P7qV5atDCSDEJ+U1lti3nQKBgQCalRNTLWLKnwKlsT+eaF9O\nsjm9Z7SWGUEFBXbZgSJculuIVLgtPghXdG2kKyQgj8eFaMmwAERTrZX5jG60glt7\nowXk1hrjofV3j6khElSN+ag6/fD751Pz9yZNDm5NILE4ciDiYxzgjZIrOZyt+Zzr\nNBYF623daKhE5oha3fzb3A==\n-----END PRIVATE KEY-----\n",
+    "client_email": "gsheet-operator@discord-bot-412100.iam.gserviceaccount.com",
+    "client_id": "114927985012267012047",
+    "auth_uri": "https://accounts.google.com/o/oauth2/auth",
+    "token_uri": "https://oauth2.googleapis.com/token",
+    "auth_provider_x509_cert_url": "https://www.googleapis.com/oauth2/v1/certs",
+    "client_x509_cert_url": "https://www.googleapis.com/robot/v1/metadata/x509/gsheet-operator%40discord-bot-412100.iam.gserviceaccount.com",
+    "universe_domain": "googleapis.com"
+  }
