+++ index.js
@@ -1,20 +1,10 @@
 // Require the necessary discord.js classes
+const client = require('./client.config');
 const fs = require('node:fs');
 const path = require('node:path');
 const { Client, Collection, Events, GatewayIntentBits} = require('discord.js');
-const { token, channelId } = require('./config.json');
-const axios = require('axios');
-const express = require('express');
-const cors = require('cors');
-const bodyParser = require('body-parser');
+const { token } = require('./config.json');
 // Create a new client instance
-const client = new Client({
-	intents:
-		[GatewayIntentBits.Guilds,
-		GatewayIntentBits.GuildMessages,
-		GatewayIntentBits.MessageContent,
-		GatewayIntentBits.GuildMembers,]
-});
 
 // When the client is ready, run this code (only once).
 // The distinction between `client: Client<boolean>` and `readyClient: Client<true>` is important for TypeScript developers.
@@ -22,13 +12,6 @@ const client = new Client({
 client.once(Events.ClientReady, readyClient => {
 	console.log(`Ready! Logged in as ${readyClient.user.tag}`);
 });
-
-client.on("messageCreate", msg => {
-	console.log(msg.content);
-	if (msg.content === "ping") {
-		msg.reply({ content: 'pong' })
-	}
-})
 
 client.commands = new Collection();
 //command handling, loading all command files and saving command paths in commandFiles 
@@ -77,40 +60,5 @@ client.on(Events.InteractionCreate, async interaction => {
 // Log in to Discord with your client's token
 client.login(token);
 
-const app = express();
-const port = 3000;
-app.use(bodyParser.json());
-app.use(cors());
-app.get('/', (req, res) => { //used for testing if the server is running
-	res.send({ msg: "server is running" });
-});
-
-app.post('/', (req, res) => {
-	const payload = req.body;
-	if (!payload.pusher) { //Only when a ciommit is pushed
-		res.status(200);
-		return;
-	}
-
-	const pusher = payload.pusher.name;
-	const repoName = payload.repository.name;
-	const commitMsg = payload.head_commit.message;
-	const commitUrl = payload.head_commit.url;
-
-	const discordMessage = `${pusher} has pushed to ${repoName} with commit: ${commitMsg}, ${commitUrl}`;
-
-	const channel = client.channels.cache.get(channelId);
-	if (channel) {
-		channel.send(discordMessage);
-	} else {
-		console.log('Channel not found');
-	}
-
-	res.status(200).send({ msg: "webhook received" })
-})
-
-app.listen(port, () => {
-	console.log(`Example app listening at http://localhost:${port}`);
-});
